:arrow_double_up: Intitial extractors and performance analysis
