+++ autoflex/types/structures.py
@@ -1,0 +1,9 @@
+from autoflex.types.core import AutoflexBaseModel
+from autoflex.types.descriptors import AutoflexParameterTypes
+
+
+class ParameterTable(AutoflexBaseModel):
+    name: str
+    types: str
+    description: str
+    special_parameters: AutoflexParameterTypes
