+++ demo/basic_class.py
@@ -25,5 +25,3 @@ class BasicClass(pydantic.BaseModel):
         This is my static method.
         """
         pass
-
-# documenter = some_method(BasicClass)
