+++ docs/api/descriptors.rst
@@ -1,0 +1,2 @@
+Descriptors
+------------
