+++ docs/api/index.rst
@@ -1,0 +1,6 @@
+API
+====
+
+.. toctree::
+
+    descriptors
