--- docs/directives/flexsummary.rst
@@ -1,2 +1,0 @@
-``flexsummary``
----------------
