+++ docs/directives/index.rst
@@ -2,11 +2,9 @@ Directives
 ==========
 
 .. include:: autoflex.rst
-.. include:: flexsummary.rst
 
 .. toctree::
    :hidden:
 
    autoflex
-   flexsummary
