--- tests/test_directive.py
@@ -1,1 +1,0 @@
-
