--- tests/test_import.py
@@ -1,3 +1,0 @@
-
-def test_basic_import():
-    import autoflex
